fixed issue with YMD
